Actualizacion de archivos

+++ app/views.py
@@ -6,7 +6,7 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, redirect, render
 from django.utils import timezone
 
-from .models import Event, Rating, RatingForm, User
+from .models import Event, Rating, Rating_Form, User
 
 
 def register(request):
@@ -129,45 +129,44 @@ def event_form(request, id=None):
     )
 
 @login_required
-def detalle_evento(request, id):
+def event_rating(request, id):
     evento = get_object_or_404(Event, pk=id)
     resenas = Rating.objects.filter(evento=evento)
+    cantidad_resenas = resenas.count()
 
     try:
         resena_existente = Rating.objects.get(usuario=request.user, evento=evento)
-        form = RatingForm(instance=resena_existente)
         editando = True
     except Rating.DoesNotExist:
         resena_existente = None
-        form = RatingForm()
         editando = False
         
     if request.method == 'POST':
         if 'guardar' in request.POST:
-            form = RatingForm(request.POST, instance=resena_existente)
+            form = Rating_Form(request.POST, instance=resena_existente)
             if form.is_valid():
                 nueva_resena = form.save(commit=False)
                 nueva_resena.usuario = request.user 
                 nueva_resena.evento = evento
                 nueva_resena.save()
                 messages.success(request, "¡Tu reseña fue guardada exitosamente!")
-                form = RatingForm()
-                return redirect('event_rating', id=evento.id) # type: ignore
+            return redirect('event_rating', id=evento.id) # type: ignore
         
-        if 'eliminar' in request.POST:
-            if resena_existente:
-                resena_existente.delete()
-                messages.success(request, "¡Tu reseña fue eliminada exitosamente!")
-                return redirect('event_rating', id=evento.id) # type: ignore
+        elif 'eliminar' in request.POST and resena_existente:
+            resena_existente.delete()
+            messages.success(request, "¡Tu reseña fue eliminada exitosamente!")
+            return redirect('event_rating', id=evento.id) # type: ignore
 
-        if 'cancelar' in request.POST:
-            form = RatingForm()
+        elif 'cancelar' in request.POST:
+            form = Rating_Form()
             return redirect('event_rating', id=evento.id) #type: ignore
-                
+
+    form = Rating_Form(instance=resena_existente)            
         
     return render(request, 'app/event_rating.html', {
         'evento': evento,
         'ratings' : resenas,
         'form': form,
-        'editando': editando
+        'editando': editando,
+        'cantidad_resenas': cantidad_resenas
     })
